Implements the Result class
removes the result struct from the Struct file and turn it into a
class.
